Update script.js
The previous version only changed the background picture when page was refreshed and not when the quotes changed by the button and I have fixed it by creating a getRandomColor function that gets called every time new quote is generated. There was also an issue with the previous version where it would check the entire array to see if it has citation..etc and not the random one that has been generated from the main array. I have fixed that and now it checks the random picked array to see what properties it has and if it does have one it will display it including the tag.